Eviction policies: LRU and LFU / some improvements in daemons
