+++ mictlanxrouter/_async/__init__.py
@@ -1,3 +1,3 @@
-from .peer_manager_daemon import run_async_healer
-from .replicator_daemon import run_replicator 
+from .peer_manager import run_async_healer
+from .replicator_daemon import run_data_replicator 
 from .replica_management import run_rm
