+++ mictlanxrouter/opentelemetry/__init__.py
@@ -1,0 +1,5 @@
+from opentelemetry.sdk.trace.export import ConsoleSpanExporter
+
+class NoOpSpanExporter(ConsoleSpanExporter):
+    def export(self, spans):
+        pass
