+++ peers-dev.json
@@ -28,7 +28,7 @@
             "WORKERS": "4",
             "ELASTIC": "true",
             "NODE_IP_ADDR": "0.0.0.0",
-            "PEERS": "mictlanx-peer-2:25002"
+            "PEERS": ""
         },
         "memory": 4294967296,
         "cpu_count": 2,
@@ -85,7 +85,7 @@
             "WORKERS": "4",
             "ELASTIC": "true",
             "NODE_IP_ADDR": "0.0.0.0",
-            "PEERS": "mictlanx-peer-2:25002 mictlanx-peer-0:25000"
+            "PEERS": ""
         },
         "memory": 4000000000,
         "cpu_count": 2,
